feat: Remove unnecessary fields from alerts
Removes unnecessary fields from alerts.

The `initialDescription` and `initialName` fields were not necessary in the `ProjectDeleteAlert` since the alert did not need to display any content.

The `hasContent` flag was added to the `ProjectFormAlert` to indicate whether the alert should display content. This allows the alert to be used for both creating and editing projects.